+++ examples/spb-page/testExternal.ts
@@ -1,0 +1,5 @@
+export class Alerter {
+  constructor() {
+    alert('Hello from javascript');
+  }
+}